# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bytech Electronics Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  salePrice?: number | null;
  stock: number;
  imageUrl?: string | null;
  categoryId?: number | null;
  categoryName?: string | null;
  featured: boolean;
  createdAt: string;
}

export interface CreateProductInput {
  name: string;
  description: string;
  price: number;
  salePrice?: number | null;
  stock: number;
  imageUrl?: string | null;
  categoryId?: number | null;
  featured: boolean;
}

export interface UpdateProductInput {
  name?: string;
  description?: string;
  price?: number;
  salePrice?: number | null;
  stock?: number;
  imageUrl?: string | null;
  categoryId?: number | null;
  featured?: boolean;
}

export interface Category {
  id: number;
  name: string;
  description?: string | null;
  productCount: number;
}

export interface CreateCategoryInput {
  name: string;
  description?: string | null;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  totalAmount: number;
  status: string;
  items: OrderItem[];
  createdAt: string;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
}

export interface CreateOrderInput {
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  items: OrderItemInput[];
}

export type UpdateOrderStatusInputStatus =
  (typeof UpdateOrderStatusInputStatus)[keyof typeof UpdateOrderStatusInputStatus];

export const UpdateOrderStatusInputStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusInput {
  status: UpdateOrderStatusInputStatus;
}

export interface Sale {
  id: number;
  title: string;
  description?: string | null;
  discountPercent: number;
  productId?: number | null;
  categoryId?: number | null;
  startDate: string;
  endDate: string;
  active: boolean;
}

export interface CreateSaleInput {
  title: string;
  description?: string | null;
  discountPercent: number;
  productId?: number | null;
  categoryId?: number | null;
  startDate: string;
  endDate: string;
}

export interface OwnerLoginInput {
  password: string;
}

export interface AuthResponse {
  success: boolean;
  role: string;
}

export interface SessionInfo {
  isOwner: boolean;
  role: string;
}

export interface StoreStats {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
  lowStockProducts: number;
  activeSales: number;
}

export type ListProductsParams = {
  categoryId?: number;
  search?: string;
  onSale?: boolean;
};