# lib/api-zod/src/generated/types/createSaleInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bytech Electronics Store API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSaleInput {
  title: string;
  description?: string | null;
  discountPercent: number;
  productId?: number | null;
  categoryId?: number | null;
  startDate: string;
  endDate: string;
}