# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bytech Electronics Store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  totalAmount: number;
  status: string;
  items: OrderItem[];
  createdAt: string;
}