# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bytech Electronics Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  salePrice?: number | null;
  stock: number;
  imageUrl?: string | null;
  categoryId?: number | null;
  categoryName?: string | null;
  featured: boolean;
  createdAt: string;
}