# lib/api-zod/src/generated/types/sessionInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bytech Electronics Store API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInfo {
  isOwner: boolean;
  role: string;
}